feat: upsert budget

+++ internal/domain/entity/entity_gen.go
@@ -29,7 +29,7 @@ type Transaction struct {
 	ExternalID    string     `json:"external_id,omitempty"`
 	Name          string     `json:"name,omitempty"`
 	Description   *string    `json:"description,omitempty"`
-	Amount        int64      `json:"amount,omitempty"`
+	Amount        float64    `json:"amount,omitempty"`
 	PaymentMethod string     `json:"payment_method,omitempty"`
 	IsIgnored     bool       `json:"is_ignored,omitempty"`
 	Date          time.Time  `json:"date,omitempty"`
@@ -44,9 +44,9 @@ type Investment struct {
 	ID         uuid.UUID `json:"id,omitempty"`
 	ExternalID string    `json:"external_id,omitempty"`
 	Name       string    `json:"name,omitempty"`
-	Amount     int64     `json:"amount,omitempty"`
+	Amount     float64   `json:"amount,omitempty"`
 	Type       string    `json:"type,omitempty"`
-	Rate       int64     `json:"rate,omitempty"`
+	Rate       float64   `json:"rate,omitempty"`
 	RateType   string    `json:"rateType,omitempty"`
 	CreatedAt  time.Time `json:"created_at,omitempty"`
 	UpdatedAt  time.Time `json:"updated_at,omitempty"`
@@ -65,7 +65,7 @@ type Account struct {
 	ID            uuid.UUID `json:"id,omitempty"`
 	ExternalID    string    `json:"external_id,omitempty"`
 	Name          string    `json:"name,omitempty"`
-	Balance       int64     `json:"balance,omitempty"`
+	Balance       float64   `json:"balance,omitempty"`
 	Type          string    `json:"type,omitempty"`
 	CreatedAt     time.Time `json:"created_at,omitempty"`
 	UpdatedAt     time.Time `json:"updated_at,omitempty"`
@@ -77,8 +77,8 @@ type CreditCard struct {
 	ID             uuid.UUID `json:"id,omitempty"`
 	Level          string    `json:"level,omitempty"`
 	Brand          string    `json:"brand,omitempty"`
-	Limit          int64     `json:"limit,omitempty"`
-	AvailableLimit int64     `json:"available_limit,omitempty"`
+	Limit          float64   `json:"limit,omitempty"`
+	AvailableLimit float64   `json:"available_limit,omitempty"`
 	CreatedAt      time.Time `json:"created_at,omitempty"`
 	UpdatedAt      time.Time `json:"updated_at,omitempty"`
 	AccountID      uuid.UUID `json:"account_id,omitempty"`
@@ -95,7 +95,7 @@ type Institution struct {
 
 type Budget struct {
 	ID        uuid.UUID `json:"id,omitempty"`
-	Amount    int64     `json:"amount,omitempty"`
+	Amount    float64   `json:"amount,omitempty"`
 	CreatedAt time.Time `json:"created_at,omitempty"`
 	UpdatedAt time.Time `json:"updated_at,omitempty"`
 	UserID    uuid.UUID `json:"user_id,omitempty"`
@@ -103,7 +103,7 @@ type Budget struct {
 
 type BudgetCategory struct {
 	ID         uuid.UUID `json:"id,omitempty"`
-	Amount     int64     `json:"amount,omitempty"`
+	Amount     float64   `json:"amount,omitempty"`
 	CreatedAt  time.Time `json:"created_at,omitempty"`
 	UpdatedAt  time.Time `json:"updated_at,omitempty"`
 	BudgetID   uuid.UUID `json:"budget_id,omitempty"`
